reconstruct codebase move all pages folder inside pages folder
