+++ src/Pages/Collector/InputComponents/Skill.js
