+++ src/Pages/DownloadResume/DownloadResume.js
@@ -1,15 +1,15 @@
 import { Container, Stack, IconButton, Box, Typography, Button, Dialog, DialogTitle, InputAdornment, TextField, Snackbar, Alert } from "@mui/material";
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import { useParams, useNavigate, useLocation } from "react-router-dom";
-import { SIMPLE_ID, MATRIX_ID, DARK_RED_MATRIX_ID, DARK_RED_SIMPLE_ID } from "../helper/Strings";
-import Simple from "../ResumeTemplates/Simple";
-import Matrix from "../ResumeTemplates/Matrix";
+import { SIMPLE_ID, MATRIX_ID, DARK_RED_MATRIX_ID, DARK_RED_SIMPLE_ID } from "../../helper/Strings";
+import Simple from "../../ResumeTemplates/Simple";
+import Matrix from "../../ResumeTemplates/Matrix";
 import { PDFViewer } from "@react-pdf/renderer";
 import { connect } from 'react-redux'
 import { Report } from "@mui/icons-material";
 import { useState } from "react";
-import { CreateOrSaveResume } from "../helper/API/Resume";
-import { setData } from '../helper/UtilityMethods';
+import { CreateOrSaveResume } from "../../helper/API/Resume";
+import { setData } from '../../helper/UtilityMethods';
 
 const BLACK = "#000";
 const DARK_RED = "#e44747";
